move encounter-list to visits-table
