+++ packages/esm-patient-chart-app/src/visit/visits-widget/past-visits-components/visits-table/index.ts
@@ -1,0 +1,3 @@
+import VisitsTable from './visits-table.component';
+
+export default VisitsTable;
