+++ packages/esm-patient-chart-app/src/visit/visits-widget/past-visits-components/visits-table/visits-table.scss
@@ -1,7 +1,6 @@
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
 @import '~@openmrs/esm-styleguide/src/vars';
-@import "../../../root.scss";
 
 .tableContainer {
   padding: 0;
@@ -48,11 +47,9 @@
   max-width: 16rem;
 }
 
-.tableHeader {
-  > :global(.cds--table-header-label) {
-    @extend .productiveHeading01;
-    color: $text-02;
-  }
+.tableHeader > :global(.cds--table-header-label) {
+  @include type.type-style('productive-heading-01');
+  color: $text-02;
 }
 
 .menuItem {
