+++ packages/esm-patient-chart-app/translations/en.json
@@ -6,6 +6,7 @@
   "address": "Address",
   "all": "All",
   "allEncounters": "All encounters",
+  "allVisits": "All visits",
   "cancel": "Cancel",
   "cancelActiveVisit": "Cancel active visit",
   "cancelVisit": "Cancel Visit",
